+++ enki/impl/cstr.hpp
